cmd/test: add OTE infrastructure and migrate tests to Ginkgo format
This commit introduces the OpenShift Tests Extension framework infrastructure
and migrates E2E tests to Ginkgo format for Secondary Scheduler Operator.

Infrastructure changes:
- Update README.md Tests section to follow OTE conventions with examples
- Update test suite path from /all to /operator/serial
- Refactor cmd/secondary-scheduler-operator-tests-ext/main.go with proper
  error handling and OTE test suite configuration
- Define the operator/serial suite with [Operator] and [Serial] tags
- Update Dockerfile.rhel7 to build and gzip the test binary
- Add .gitignore entries for OTE test binaries

Test migration changes:
- Add Ginkgo specs to test/e2e/operator.go for operator functionality tests
- Create test/e2e/ha_mode.go with test functions and Ginkgo specs for HA tests
- Simplify test/e2e/ha_mode_test.go to minimal legacy Go test entry point
- Tests tagged with [sig-scheduling][Operator][Serial] for OTE framework
- Maintain backward compatibility with legacy Go tests
