+++ cmd/secondary-scheduler-operator-tests-ext/main.go
@@ -1,34 +1,36 @@
-/*
-This command is used to run the Secondary Scheduler Operator tests extension for OpenShift.
-It registers the Secondary Scheduler  Operator tests with the OpenShift Tests Extension framework
-and provides a command-line interface to execute them.
-For further information, please refer to the documentation here
-https://github.com/openshift-eng/openshift-tests-extension/blob/main/cmd/example-tests/main.go
-*/
 package main
 
 import (
 	"context"
+	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
 	"k8s.io/component-base/cli"
+	"k8s.io/klog/v2"
 
 	otecmd "github.com/openshift-eng/openshift-tests-extension/pkg/cmd"
 	oteextension "github.com/openshift-eng/openshift-tests-extension/pkg/extension"
+	oteginkgo "github.com/openshift-eng/openshift-tests-extension/pkg/ginkgo"
 	"github.com/openshift/secondary-scheduler-operator/pkg/version"
 
-	"k8s.io/klog/v2"
+	_ "github.com/openshift/secondary-scheduler-operator/test/e2e"
 )
 
 func main() {
-	command := newOperatorTestCommand(context.Background())
+	command, err := newOperatorTestCommand(context.Background())
+	if err != nil {
+		klog.Fatal(err)
+	}
 	code := cli.Run(command)
 	os.Exit(code)
 }
 
-func newOperatorTestCommand(ctx context.Context) *cobra.Command {
-	registry := prepareOperatorTestsRegistry()
+func newOperatorTestCommand(ctx context.Context) (*cobra.Command, error) {
+	registry, err := prepareOperatorTestsRegistry()
+	if err != nil {
+		return nil, fmt.Errorf("failed to prepare operator tests registry: %w", err)
+	}
 
 	cmd := &cobra.Command{
 		Use:   "secondary-scheduler-operator-tests-ext",
@@ -49,7 +51,7 @@ func newOperatorTestCommand(ctx context.Context) *cobra.Command {
 
 	cmd.AddCommand(otecmd.DefaultExtensionCommands(registry)...)
 
-	return cmd
+	return cmd, nil
 }
 
 // prepareOperatorTestsRegistry creates the OTE registry for this operator.
@@ -57,10 +59,27 @@ func newOperatorTestCommand(ctx context.Context) *cobra.Command {
 // Note:
 //
 // This method must be called before adding the registry to the OTE framework.
-func prepareOperatorTestsRegistry() *oteextension.Registry {
+func prepareOperatorTestsRegistry() (*oteextension.Registry, error) {
 	registry := oteextension.NewRegistry()
 	extension := oteextension.NewExtension("openshift", "payload", "secondary-scheduler-operator")
 
+	// The following suite runs tests that verify the operator's behaviour.
+	// This suite is executed only on pull requests targeting this repository.
+	// Tests tagged with both [Operator] and [Serial] are included in this suite.
+	extension.AddSuite(oteextension.Suite{
+		Name:        "openshift/secondary-scheduler-operator/operator/serial",
+		Parallelism: 1,
+		Qualifiers: []string{
+			`name.contains("[Operator]") && name.contains("[Serial]")`,
+		},
+	})
+
+	specs, err := oteginkgo.BuildExtensionTestSpecsFromOpenShiftGinkgoSuite()
+	if err != nil {
+		return nil, fmt.Errorf("couldn't build extension test specs from ginkgo: %w", err)
+	}
+
+	extension.AddSpecs(specs)
 	registry.Register(extension)
-	return registry
+	return registry, nil
 }
